refactor(cart): 장바구니 관련 로직 유틸로 분리
